Code cleanup and fix graph deletion bug
